Created companion RSS reader for my new blog
I created a new blog to write about my tech projects and to
house my writing about contemporary art. The blog resides at
http://lisawilliams.github.io/lisa/

Inspired by http://scripting.com/river.html, I wanted to make
a companion RSS reader for my blog.

I used TinyTinyRSS and some helpful directions I found on how
to host the feed reader on Heroku, and now have a working instance
at the following URL:

https://my-fancy-ttrss.herokuapp.com/

The instructions for setting up TTRSS on Heroku are here:
https://github.com/serl/ttrss-heroku

I set up a scheduled job to update the feeds as indicated in the instructions,
but it's not clear to me that the scheduled job is working.

Visiting Heroku, and clicking on the dyno for this app and then on the
Console tab opens a command line for the Heroku dyno for the app.
Running `update` at that command line does update the feeds,
so in the worst case scenario feeds can be updated manually.

Responds to the following issue: https://github.com/lisawilliams/lisa/issues/20